+++ src/dipdup/indexes/evm_subsquid_events/matcher.py
@@ -85,21 +85,17 @@ def match_events(
     topics: dict[str, dict[str, str]],
 ) -> deque[MatchedEventsT]:
     """Try to match contract events with all index handlers."""
-    # FIXME: Don't do it every time
-    topics_cache = {f'{h.contract.module_name}:{h.name}': topics[h.contract.module_name][h.name] for h in handlers}
     matched_handlers: deque[MatchedEventsT] = deque()
-    events = deque(events)
 
-    for event in events.copy():
+    for event in events:
         for handler_config in handlers:
-            topic = topics_cache[f'{handler_config.contract.module_name}:{handler_config.name}']
-            if topic != event.topics[0]:
+            typename = handler_config.contract.module_name
+            name = handler_config.name
+            if topics[typename][name] != event.topics[0]:
                 continue
 
             arg = prepare_event_handler_args(package, handler_config, event)
             matched_handlers.append((handler_config, arg))
-
-            events.remove(event)
             break
 
     return matched_handlers